// src/types/BigCommercePermission.ts
export enum BigCommercePermission {
    All = "all",
    Content = "store_v2_content",
    ContentRead = "store_v2_content_read_only",
    CheckoutContent = "store_content_checkout",
    CheckoutContentRead = "store_content_checkout_read_only",
    Customers = "store_v2_customers",
    CustomersRead = "store_v2_customers_read_only",
    CustomersLogin = "store_v2_customers_login",
    InformationSettings = "store_v2_information",
    InformationSettingsRead = "store_v2_information_read_only",
    Marketing = "store_v2_marketing",
    MarketingRead = "store_v2_marketing_read_only",
    Orders = "store_v2_orders",
    OrdersRead = "store_v2_orders_read_only",
    OrderTransactions = "store_v2_order_transactions",
    OrderTransactionsRead = "store_v2_order_transactions_read_only",
    CreatePayments = "store_payments_access_token_create",
    GetPaymentMethodsRead = "store_payments_methods_read",
    Products = "store_v2_products",
    ProductsRead = "store_v2_products_read_only",
    Themes = "store_themes_manage",
    ThemesRead = "store_themes_read_only",
    Carts = "store_cart",
    CartsRead = "store_cart_read_only",
    Checkouts = "store_checkout",
    CheckoutsRead = "store_checkout_read_only",
    SitesRoutes = "store_sites",
    SitesRoutesRead = "store_sites_read_only",
    ChannelSettingsd = "store_channel_settings",
    ChannelSettingsRead = "store_channel_settings_read_only",
    ChannelListings = "store_channel_listings",
    ChannelListingsRead = "store_channel_listings_read_only",
    StorefrontApiTokens = "store_storefront_api",
    StorefrontApiCustomerImpersonationTokens = "store_storefront_api_customer_impersonation"
}